requirements.txt
# Local dev: use editable core
course-scheduler-core @ git+https://github.com/scorvachoa/course-scheduler-uc-core@v0.1.3

fastapi
uvicorn
reportlab

